rename download_api sub modules
